mudando chave totalPrice da tabela sales para decimal

+++ back-end/src/database/migrations/20230120190558-create-sale.js
@@ -25,7 +25,7 @@ module.exports = {
         },
       },
       totalPrice: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(9, 2),
         field: "total_price",
       },
       deliveryAddress: {
